# documentatie/openAPI.yaml
openapi: 3.0.0
info:
  title: Opdrachtenbox API
  description: API voor het beheren van opdrachten voor studenten van de opleiding CMD
  version: 0.0.3
servers:
  - url: http://localhost:8080/
    description: Local development server
...
paths:
  /registrations:
    patch:
      tags:
        - Aanmeldingen
      summary: voor het accepteren of afwijzen van een aanmelding
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: integer
                  format: int32
                  example: 12
                userId:
                  type: integer
                  format: int32
                  example: 312
                accepted:
                  type: boolean
                  example: true
                response:
                  type: string
                  maxLength: 400
                  example: Welkom bij het project
              required:
                - accepted
      description: "Alleen toegankelijk voor: de opdrachtgever zelf"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: aanmelding succesvol geaccepteerd of afgewezen
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /registrations/existing-user-registrations:
    get:
      tags:
        - Aanmeldingen
      summary: ophalen van alle taskIds waar de student zich voor heeft aangemeld
      description: "Alleen toegankelijk voor: student"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: aanmeldingen succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
                  example: 12
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de student rol
        "5XX":
          $ref: "#/components/responses/ServerError"
  /registrations/{taskId}:
    get:
      tags:
        - Aanmeldingen
      summary: ophalen van alle aanmeldingen voor de opdrachtgever die is ingelogd
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          description: de id van het project
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: aanmeldingen succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetRegistrationDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Aanmeldingen
      summary: nieuwe aanmelding toevoegen
      parameters:
        - name: taskId
          in: path
          required: true
          description: de id van de task waar een aanmelding naar gedaan wordt
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/text:
            schema:
              type: string
              example: ik ben goed in adobe premiere pro
              maxLength: 400
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: aanmelding succesvol toegevoegd
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de student rol
        "404":
          description: taskId is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
...
  /projects:
    get:
      tags:
        - Projecten
      summary: alle projecten ophalen
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      parameters:
        - in: query
          name: businessId
          schema:
            type: integer
            format: int32
          required: false
      responses:
        "200":
          description: projecten succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetProjectDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Projecten
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 50
                  example: Nieuwe project titel
                description:
                  type: string
                  maxLength: 400
                  example: Nieuwe project beschrijving
                image:
                  type: string
                  format: binary
                  description: afbeelding van het project
      summary: nieuw project aanmaken
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: project succesvol aangemaakt
          content:
            application/json:
              schema:
                type: integer
                format: int32
                example: 25
                description: Het id van het aangemaakte project
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "5XX":
          $ref: "#/components/responses/ServerError"
...
  /projects/{projectId}:
    get:
      tags:
        - Projecten
      summary: Een enkel project ophalen
      parameters:
        - name: projectId
          in: path
          required: true
          description: de id van het project
          schema:
            type: integer
            format: int32
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: Project succesvol opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProjectDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "404":
          description: project heeft een foute opbouw
        "5XX":
          $ref: "#/components/responses/ServerError"
  /skills:
    get:
      tags:
        - Skills
      summary: ophalen van alle skills
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: skills succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSkillDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Skills
      summary: nieuwe skill toevoegen
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      requestBody:
        content:
          application/text:
            schema:
              type: string
              maxLength: 50
              example: adobe premiere pro
      responses:
        "201":
          description: skill succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSkillDto"
        "400":
          description: skill heeft een foute opbouw, of bestaat al
        "401":
          description: gebruiker is niet ingelogd
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "5XX":
          description: server error
  /skills/{skillId}/acceptance:
    patch:
      tags:
        - Skills
      summary: Nieuwe skills accepteren of afwijzen
      description: "Alleen toegankelijk voor: docenten"
      security:
        - sessionCookieAuth: []
      parameters:
        - name: skillId
          in: path
          required: true
          description: de id van de skill
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: boolean
              example: true
      responses:
        "200":
          description: Skill succesvol geaccepteerd of afgewezen
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de docenten rol
        "404":
          description: skillId is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /skills/{skillId}/name:
    patch:
      tags:
        - Skills
      summary: skill naam aanpassen
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      parameters:
        - name: skillId
          in: path
          required: true
          description: de id van de skill
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/text:
            schema:
              type: string
              maxLength: 50
              example: adobe premiere pro
      responses:
        "200":
          description: skill naam succesvol aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "404":
          description: skillId is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  responses:
    BadRequest:
      description: request body heeft een foute opbouw
    UnAuthenticated:
      description: gebruiker is niet ingelogd
    ServerError:
      description: server fout
  schemas:
    GetProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        title:
          type: string
          maxLength: 50
          example: Project titel
        description:
          type: string
          maxLength: 400
          example: Project beschrijving
        projectTopSkills:
          type: array
          items:
            $ref: "#/components/schemas/GetSkillDto"
        business:
          $ref: "#/components/schemas/BusinessDto"
        photo:
          $ref: "#/components/schemas/FileDto"
      required:
        - projectId
        - title
        - description
        - business
        - projectTopSkills
...
    GetSkillDto:
      description: een skill object
      type: object
      properties:
        skillId:
          type: integer
          format: int32
          example: 20
        name:
          type: string
          maxLength: 50
          example: adobe premiere pro
        isPending:
          type: boolean
          example: false
      required:
        - skillId
        - name
        - isPending
    GetRegistrationDto:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
          example: 43
        reason:
          type: string
          example: Ik wil mij aanmelden omdat...
          description: Reden van aanmelding, gegeven door de student
          maxLength: 400
        accepted:
          type: boolean
        response:
          type: string
          example: Jij bent geschikt voor dit project
          description: Reactie van de opdrachtgever
        student:
          $ref: "#/components/schemas/StudentDto"
      required:
        - id
        - reason
        - student
    FileDto:
      type: object
      properties:
        path:
          type: string
          example: /image.png
          maxLength: 255
...
    StudentDto:
      type: object
      properties:
        userId:
          type: integer
          format: int32
          example: 312
        username:
          type: string
          maxLength: 50
          example: mijn naam
        description:
          type: string
          maxLength: 400
          example: ik ben een student met een passie voor ...
        profilePicture:
          $ref: "#/components/schemas/FileDto"
        email:
          type: string
          example: name@student.han.nl
        cv:
          $ref: "#/components/schemas/FileDto"
        skills:
          type: array
          items:
            type: object
            properties:
              skill:
                $ref: "#/components/schemas/GetSkillDto"
              description:
                type: string
                maxLength: 400
                example: ik ben goed in adobe premiere pro
      required:
        - username
        - description
        - profilePicture
        - cv
        - skills
    BusinessDto:
      type: object
      properties:
        businessId:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          maxLength: 255
          example: mijn naam
        description:
          type: string
          maxLength: 4000
          example: Dit bedrijf is een bedrijf dat ...
        photo:
          $ref: "#/components/schemas/FileDto"
        location:
          type: string
          maxLength: 255
      required:
        - name
        - description
        - photo
        - location
...